Moved account-related queries to accountQueries.go
